#12 merged slidecard to cardslide.  added a link practice on navbar
